WIP: multiple consumer fsms  manage connections
Rather than dependning on shovel, start multiple consumers, each
pointing to a different remote rabbit node. OR NOT! This refactor takes
a list of `[{Count::pos_integer(), amqp_uti()}]` pairs and starts
`Count` consumers for each `amqp_uri()`. The `amqp_uri()` can be a local
rabbitmq that is shovelled too, or a remote one, whatever.

The `rabl_consumer_fsm`s primary role is not to crash the
app. Eventually rabl will be started as an included app for riak, and
not rablicating should never crash the database node. The consumer
manages a connection to rabbit. If that connection ever breaks, the
consumer attempts to reconnect (with back off) for a configured number
of times, then stops. The `rabl_sup` will restart it, and off it goes
again.

Likewise, if the consumer gets a weird message, it logs and keeps
ticking. If it fails to put in riak, it nacks, and keeps ticking.

TODO add some back off on overload of riak.